moved inspect_string_multiple_rules into separate file
