+++ dlp/snippets/custom_infotype_test.py
@@ -61,43 +61,3 @@ def test_inspect_string_without_overlap(capsys: pytest.LogCaptureFixture) -> Non
     out, _ = capsys.readouterr()
     assert "example.com" in out
     assert "example.org" not in out
-
-
-def test_inspect_string_multiple_rules_patient(
-    capsys: pytest.LogCaptureFixture,
-) -> None:
-    custom_infotype.inspect_string_multiple_rules(
-        GCLOUD_PROJECT, "patient name: Jane Doe"
-    )
-
-    out, _ = capsys.readouterr()
-    assert "Likelihood: 4" in out
-
-
-def test_inspect_string_multiple_rules_doctor(capsys: pytest.LogCaptureFixture) -> None:
-    custom_infotype.inspect_string_multiple_rules(GCLOUD_PROJECT, "doctor: Jane Doe")
-
-    out, _ = capsys.readouterr()
-    assert "No findings" in out
-
-
-def test_inspect_string_multiple_rules_quasimodo(
-    capsys: pytest.LogCaptureFixture,
-) -> None:
-    custom_infotype.inspect_string_multiple_rules(
-        GCLOUD_PROJECT, "patient name: quasimodo"
-    )
-
-    out, _ = capsys.readouterr()
-    assert "No findings" in out
-
-
-def test_inspect_string_multiple_rules_redacted(
-    capsys: pytest.LogCaptureFixture,
-) -> None:
-    custom_infotype.inspect_string_multiple_rules(
-        GCLOUD_PROJECT, "name of patient: REDACTED"
-    )
-
-    out, _ = capsys.readouterr()
-    assert "No findings" in out
